Added section to explain menus in extensions
As yet, there is no documentation that explains how to manipulate Platform's menu system in the context of an extension. This should serve as a jumping-off point.

I felt it particularly important to explain the menu functionality because it's not a direct extension of Laravel's underlying functionality, as is the case with providers, routes, and some of the other functionality that is inherent to extensions.

It would be ideal to add a couple more 5th-level sections that explain the other less straightforward sections in `extension.php`.